# Conservatively remove `InitializationStore#newFieldValueAfterMethodCall` to optimize performance