--- a/checker/src/main/java/org/checkerframework/checker/initialization/InitializationStore.java
+++ b/checker/src/main/java/org/checkerframework/checker/initialization/InitializationStore.java
@@ -1,7 +1,5 @@
 package org.checkerframework.checker.initialization;
 
-import org.checkerframework.checker.initialization.qual.Initialized;
-import org.checkerframework.checker.initialization.qual.NotOnlyInitialized;
 import org.checkerframework.dataflow.cfg.visualize.CFGVisualizer;
 import org.checkerframework.dataflow.expression.ClassName;
 import org.checkerframework.dataflow.expression.FieldAccess;
@@ -9,17 +7,13 @@ import org.checkerframework.dataflow.expression.JavaExpression;
 import org.checkerframework.dataflow.expression.ThisReference;
 import org.checkerframework.framework.flow.CFAbstractStore;
 import org.checkerframework.framework.flow.CFValue;
-import org.checkerframework.framework.type.AnnotatedTypeMirror;
-import org.checkerframework.framework.type.GenericAnnotatedTypeFactory;
 import org.plumelib.util.ToStringComparator;
 
 import java.util.HashSet;
 import java.util.Set;
 
-import javax.lang.model.element.AnnotationMirror;
 import javax.lang.model.element.Element;
 import javax.lang.model.element.VariableElement;
-import javax.lang.model.type.TypeKind;
 
 /**
  * A store that extends {@code CFAbstractStore} and additionally tracks which fields of the 'self'
@@ -133,6 +127,9 @@ public class InitializationStore extends CFAbstractStore<CFValue, Initialization
         return result;
     }
 
+    /*
+     * TODO: implement a meaningful `isDeclaredInitialized`.
+     *
     @Override
     protected CFValue newFieldValueAfterMethodCall(
             FieldAccess fieldAccess,
@@ -144,19 +141,22 @@ public class InitializationStore extends CFAbstractStore<CFValue, Initialization
 
         return super.newFieldValueAfterMethodCall(fieldAccess, atypeFactory, value);
     }
+    */
 
-    /**
+    /*
      * Determine whether the given field is declared as {@link Initialized} (taking into account
      * viewpoint adaption for {@link NotOnlyInitialized}).
      *
      * @param fieldAccess the field to check
      * @return whether the given field is declared as {@link Initialized} (taking into account
      *     viewpoint adaption for {@link NotOnlyInitialized})
-     */
+     *
     protected boolean isDeclaredInitialized(FieldAccess fieldAccess) {
+        // Returning false is the conservative answer, but not much faster than asking the ATF.
+        return false;
+        /*
         InitializationParentAnnotatedTypeFactory atypeFactory =
                 (InitializationParentAnnotatedTypeFactory) analysis.getTypeFactory();
-
         AnnotatedTypeMirror declField = atypeFactory.getAnnotatedType(fieldAccess.getField());
         if (!declField.hasAnnotation(atypeFactory.INITIALIZED)) {
             return false;
@@ -189,6 +189,7 @@ public class InitializationStore extends CFAbstractStore<CFValue, Initialization
             return true;
         }
     }
+    */
 
     @Override
     protected String internalVisualize(CFGVisualizer<CFValue, InitializationStore, ?> viz) {