@wmdietl 2024-02-20
@flo2702 I indeed confused myself with the different versions of the benchmark. I now added `Issue1438c.java` that tests instance fields accessed from an instance method.
There is a significant speedup for this common case.
Please let me know what you think of this change.