@wmdietl 2024-03-01
@flo2702 Thanks for looking into this!
I tried your `fieldAccess.getField().getAnnotation(Initialized.class) != null` proposal: it didn't change timing and I think it always returns false.
So I went with the conservative answer to just return false and simply commented out this logic.
Final timing for me:

Issue1438.java: ~19s
Issue1438b.java: ~55s
Issue1438c.java: ~12s

I've filed #720  to add more tests for `NotOnlyInitialized` and see whether this is too conservative.
#719 for additional performance improvements.